[Misc] Update to UE 5.1.0

+++ Source/DogFight/Public/DogFightTypes.h
@@ -36,8 +36,8 @@ struct FCameraFocusEvent
 	GENERATED_BODY();
 
 	int32 OwnerPlayerId;
-	float LocationX;
-	float LocationY;
+	double LocationX;
+	double LocationY;
 
 	ECameraFocusEventType::Type EventType;
 };
